+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+* Material buttons will now only re-render itself if it's been selected or needs to be deselected
 
 ## [1.5.2] - December 22, 2021
 ### Added